requirements.txt
pandas
numpy
scipy
requests
python-dotenv
matplotlib